Expand on HTTP/3 extension aspects
By switching from Frames to special signal values at byte position 0,
we are really extending HTTP/3 rules. So this is an attempt to paint
that out more cleanly and tidy up some loose ends or stale text in the
draft.

The outcome is that we make it clear that WebTransport streams never use
frames. Instead, they are always a sequence of some preamble at the
start of a stream, followed by a session ID, followed by user-specified
stream data. This makes it more obvious that endpoints should never send
0x41 such that the peer could interpret it as being a frame on any
frame-bearing stream (like the control stream). If you understand
WebTransport but didn't enable it, then receiving 0x41 is also bad, so
blow up.