Tweak how HTTP/3 error codes are presented for use
This change is editorial. It is not intended to redefine how the error
codes operate but instead clarify their usage and make them more
consistent with how HTTP/3 and QUIC do things.

So, we drop `H3` prefix from the error codes, because this is all about
WebTransport.

Then we make a single label - WEBTRANSPORT_APPLICATION_ERROR - that
covers the entire range of codes that we care about. This is similar to
how QUIC's CRYPTO_ERROR works. And in the part we talk about mapping,
add a little callout to the label similar to https://datatracker.ietf.org/doc/html/rfc9001#section-4.8

Fixes #117